indexes created on the columns: to_wallet_id (table transactions), from_wallet_id (table transactions)
